Use long to marshall unsigned int since WebKit bridge can't deal with uint.
svn path=/branches/qtcpp/qtproj/weblobby/; revision=538

+++ unitsynchandler.cpp
@@ -701,7 +701,7 @@ void UnitsyncHandler::removeAllArchives() {
     logger.debug("call RemoveAllArchives(", ")");
     return fptr_RemoveAllArchives();
 }
-unsigned int UnitsyncHandler::getArchiveChecksum(QString archiveName) {
+long UnitsyncHandler::getArchiveChecksum(QString archiveName) {
     if (fptr_GetArchiveChecksum == NULL) {
         logger.error("Bad function pointer: GetArchiveChecksum");
         throw bad_fptr("GetArchiveChecksum");
@@ -893,7 +893,7 @@ QString UnitsyncHandler::getMapArchiveName(int index) {
     logger.debug("call GetMapArchiveName(", index, ")");
     return QString(fptr_GetMapArchiveName(index));
 }
-unsigned int UnitsyncHandler::getMapChecksum(int index) {
+long UnitsyncHandler::getMapChecksum(int index) {
     if (fptr_GetMapChecksum == NULL) {
         logger.error("Bad function pointer: GetMapChecksum");
         throw bad_fptr("GetMapChecksum");
@@ -901,7 +901,7 @@ unsigned int UnitsyncHandler::getMapChecksum(int index) {
     logger.debug("call GetMapChecksum(", index, ")");
     return fptr_GetMapChecksum(index);
 }
-unsigned int UnitsyncHandler::getMapChecksumFromName(QString mapName) {
+long UnitsyncHandler::getMapChecksumFromName(QString mapName) {
     if (fptr_GetMapChecksumFromName == NULL) {
         logger.error("Bad function pointer: GetMapChecksumFromName");
         throw bad_fptr("GetMapChecksumFromName");
@@ -1045,7 +1045,7 @@ long UnitsyncHandler::getPrimaryModChecksum(int index) {
     logger.debug("call GetPrimaryModChecksum(", index, ")");
     return fptr_GetPrimaryModChecksum(index);
 }
-unsigned int UnitsyncHandler::getPrimaryModChecksumFromName(QString name) {
+long UnitsyncHandler::getPrimaryModChecksumFromName(QString name) {
     if (fptr_GetPrimaryModChecksumFromName == NULL) {
         logger.error("Bad function pointer: GetPrimaryModChecksumFromName");
         throw bad_fptr("GetPrimaryModChecksumFromName");
